requirements.txt
markdown
PyYAML
requests
python-dotenv
